+++ app/src/main/java/org/simple/clinic/settings/SettingsEffectHandler.kt
@@ -9,7 +9,6 @@ import io.reactivex.ObservableTransformer
 import org.simple.clinic.appupdate.AppUpdateState
 import org.simple.clinic.appupdate.CheckAppUpdateAvailability
 import org.simple.clinic.user.UserSession
-import org.simple.clinic.user.UserSession.LogoutResult.Success
 import org.simple.clinic.util.filterAndUnwrapJust
 import org.simple.clinic.util.scheduler.SchedulersProvider