+++ app/src/main/java/org/simple/clinic/settings/SettingsUi.kt
@@ -13,4 +13,8 @@ interface SettingsUi {
   fun showAppUpdateButton()
 
   fun hideAppUpdateButton()
+
+  fun showLoggingOutProgressIndicator()
+
+  fun hideLoggingOutProgressIndicator()
 }