+++ app/src/main/java/org/simple/clinic/settings/SettingsUiRenderer.kt
@@ -8,6 +8,15 @@ class SettingsUiRenderer(private val ui: SettingsUi) : ViewRenderer<SettingsMode
     renderUserDetails(model)
     renderCurrentLanguage(model)
     renderAppVersion(model)
+    renderLoggingOutProgress(model.isUserLoggingOut)
+  }
+
+  private fun renderLoggingOutProgress(userLoggingOut: Boolean?) {
+    if (userLoggingOut == true) {
+      ui.showLoggingOutProgressIndicator()
+    } else {
+      ui.hideLoggingOutProgressIndicator()
+    }
   }
 
   private fun renderUserDetails(model: SettingsModel) {