+++ app/src/main/java/org/simple/clinic/settings/UiActions.kt
@@ -5,4 +5,5 @@ interface UiActions {
   fun openLanguageSelectionScreen()
   fun showConfirmLogoutDialog()
   fun restartApp()
+  fun goBack()
 }